move full_class and document
