+++ qcodes/instrument/channel.py
@@ -25,9 +25,9 @@ from qcodes.parameters import (
     Parameter,
 )
 from qcodes.parameters.multi_channel_instrument_parameter import InstrumentModuleType
+from qcodes.utils import full_class
 from qcodes.validators import Validator
 
-from ..utils.helpers import full_class
 from ..utils.metadata import Metadatable
 from .instrument import Instrument
 from .instrument_base import InstrumentBase
