+++ qcodes/measure.py
@@ -4,7 +4,7 @@ from typing import Optional, Sequence
 from qcodes.actions import _actions_snapshot
 from qcodes.loops import Loop
 from qcodes.parameters import Parameter
-from qcodes.utils.helpers import full_class
+from qcodes.utils import full_class
 from qcodes.utils.metadata import Metadatable
