+++ qcodes/parameters/parameter_base.py
@@ -21,8 +21,7 @@ from typing import (
     overload,
 )
 
-from qcodes.utils import DelegateAttributes, qcodes_abstractmethod
-from qcodes.utils.helpers import full_class
+from qcodes.utils import DelegateAttributes, full_class, qcodes_abstractmethod
 from qcodes.utils.metadata import Metadatable
 from qcodes.validators import Enum, Ints, Validator
