+++ qcodes/utils/full_class.py
@@ -1,0 +1,3 @@
+def full_class(obj: object) -> str:
+    """The full importable path to an object's class."""
+    return type(obj).__module__ + "." + type(obj).__name__
